personl voice chatbot/requirements.txt
SpeechRecognition
pyttsx3
wikipedia
pyaudio
customtkinter
packaging
Pillow
